third_party/WebKit/LayoutTests/http/tests/inspector/compiler-source-mapping-debug-expected.txt

CONSOLE MESSAGE: line 2: button clicked!
Tests installing compiler source map in scripts panel.



Running: testSetBreakpoint
Script source was shown.
Script execution paused.
Call stack:
    0) handleClick (source1.js:16)
    1) clickButton (compiler-source-mapping-debug.html:13)
Script execution resumed.
